explicitly cast the sprintf arg to int
git-svn-id: https://svn.parrot.org/parrot/trunk@23786 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ src/events.c
@@ -1461,7 +1461,7 @@ do_event(PARROT_INTERP, NOTNULL(parrot_event* event), NULLOK(opcode_t *next))
             pt_suspend_self_for_gc(interp);
             break;
         default:
-            fprintf(stderr, "Unhandled event type %d\n", event->type);
+            fprintf(stderr, "Unhandled event type %d\n", (int)event->type);
             break;
     }
     mem_sys_free(event);
